--- a/packages/next/build/webpack/loaders/next-babel-loader.js
+++ b/packages/next/build/webpack/loaders/next-babel-loader.js
@@ -12,6 +12,7 @@ module.exports = babelLoader.custom(babel => {
         dev: opts.dev
       }
       const loader = Object.assign({
+        cacheCompression: false,
         cacheDirectory: true
       }, opts)
       delete loader.isServer